DonorsAndDonations.tsx: Show updatedAt field as date instead of createdAt (#1691)
When a bank transaction is imported from IRIS, we receive only the date of the transaction, and not the hours. Due to that it was decided that
returned donations will be ordered by the updatedAt field, rather than createdAt.

This change ensures that, we properly show the how much time has passed since bank donation.

Cleaned up some commented code.